requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.5
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
mysql-connector-python==8.0.25
mysqlclient==2.0.3
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
